// src/shared/enums/dynamic-filter-builder.enum.ts
export enum FilterOperatorEnum {
  EQUAL = 'eq',
  ILIKE = 'ilike',
  LIKE = 'like',
  GREATER_THAN = 'gt',
  GREATER_THAN_OR_EQUAL = 'gte',
  LESS_THAN = 'lt',
  LESS_THAN_OR_EQUAL = 'lte',
  BETWEEN = 'between',
  IS_NULL = 'is_null',
  IS_NOT_NULL = 'is_not_null',
  IN = 'in',
  NOT_IN = 'notIn',
}